+++ Handcrafted_Features_Extraction/Handcrafted_Features_Extraction_Body_Tracking_Modality/handcrafted_features_extraction_body_tracking_modality/conf.py
@@ -37,6 +37,23 @@ DATA_PATH = os.path.join(DATASETS_FOLDER, CUSTOM_SETTINGS["dataset_config"]["dat
 
 DATASET = CUSTOM_SETTINGS["dataset_config"]["dataset_name"]
 
-MODALITY = CUSTOM_SETTINGS["dataset_config"].get("modality", "default_modality")
+#MODALITY = CUSTOM_SETTINGS["dataset_config"].get("modality", "default_modality")
 
-OUTPUT_MODALITY_FOLDER = os.path.join(OUTPUTS_FOLDER, DATASET, MODALITY)
+#OUTPUT_MODALITY_FOLDER = os.path.join(OUTPUTS_FOLDER, DATASET, MODALITY)
+
+if "modality" in CUSTOM_SETTINGS["dataset_config"]:
+    modality = CUSTOM_SETTINGS["dataset_config"]["modality"]
+else:
+    modality = "default_modality"
+
+MODALITY_FOLDER = os.path.join(
+    OUTPUTS_FOLDER,
+    CUSTOM_SETTINGS["dataset_config"]["dataset_name"],
+    modality,
+)
+
+EMOTION_TO_LABEL = {
+    "BORED": "01",
+    "ENGAGED": "02",
+    "FRUSTRATED": "03"
+}